Move valla components
