+++ package.json
@@ -11,7 +11,7 @@
     "start": "vite",
     "build": "vite build",
     "test": "vitest",
-    "serve": "vite preview --port 3000"
+    "preview": "vite-plugin-dotenv && vite preview"
   },
   "author": {
     "name": "Alejandro Sanchez",
