+++ src/front/js/module_valla/googlemap_vallas.jsx
