+++ src/front/js/module_valla/table_valla.jsx
